requirements.txt
click==6.7
python-decouple==3.1
Unipath==1.1
